✨ feat: manually `cherry-pick` with pyincpp.
1. add comments for pub struct
2. adjust `Deque::len`: usize -> i32 for negative index compatibility
3. add check for `Deque::push`
4. add Index/IndexMut for `Deque`
5. refine code
6. add `Dict::find`
7. adjust `Dict::remove`: Option -> bool
8. remove `partial_cmp` for Dict
9. let `Int::new()` has an empty Vec and optimize `Int`
10. adjust `List::count`: i32 -> usize
11. add `Set::find`
12. add `Set::min/max`
13. rename `Set::insert` -> `Set::add`
14. add `Set::pop`

+++ src/list.rs
@@ -5,6 +5,7 @@ use std::{
 
 use crate::utility;
 
+/// Lists are mutable sequences, typically used to store collections of homogeneous items.
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Default, Hash)]
 pub struct List<T> {
     data: Vec<T>,
@@ -48,11 +49,11 @@ impl<T> List<T> {
     }
 
     /// Counts the number of occurrence of the specified `element`.
-    pub fn count(&self, element: &T) -> i32
+    pub fn count(&self, element: &T) -> usize
     where
         T: PartialEq,
     {
-        self.data.iter().filter(|&x| x == element).count() as i32
+        self.data.iter().filter(|&x| x == element).count()
     }
 
     /// Return a list that eliminates duplicate elements and keep the original relative order of elements.
